Refactor content script into modules
