+++ manifest.json
@@ -22,7 +22,12 @@
   "content_scripts": [
     {
       "matches": [ "https://docs.google.com/document/*" ],
-      "js": [ "content_script.js" ],
+      "js": [
+        "modules/selection.js",
+        "modules/dom.js",
+        "modules/storage.js",
+        "content_script.js"
+      ],
       "css": [ "styles.css" ]
     }
   ]
